+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.25.52
 * Scroll improved
 * Attempt to fix unread counters on sync
+* Dialog force render on updateOutBox added
 
 ## 0.25.51
 * Unread counter improvements